many changes made queue add remove list funcitons added upgrade youtube downloader function added memory function added

+++ plugins/youtube.py
@@ -9,10 +9,12 @@ plgn = Plugin('youtube')
 @plgn.command('tell')
 def tell(data,what=None):
     string = """Follow the following commands :
-            download (link) a
+            download (link) [a]
                 This function lets you download a video or audio directly. 
                 'a' is optional parameter for downloading the audio file.
-            queue (link) 
+            queue
+                Returns a list of all links that are noted in download_queue.txt
+            queue add (link) [a]
                 This function appends the link to the download queue. Pass 'a' to download the audio.
             begin 
                 This initiates the download of the queue.
@@ -20,6 +22,10 @@ def tell(data,what=None):
                 This function will simply return all the items that are present in the downloads directory on raspberry pi.
             ip
                 This will return the ip of DJPI. 
+            memory
+                gives the present space usage in pi
+            update youtube downloader
+                pip install --upgrade youtube_dl
         """     
     return string
 
@@ -61,12 +67,15 @@ def download(data, what,param):
         return str(output)
     return "Done downloading "+ "\n" +"\n".join(final)
 
-@plgn.command('queue <(?P<what>[-a-zA-Z0-9 `,;!@#$%^&*()_=.{}:"\?\<\>/\[\'\]\\n]+)> *(?P<param>[aA]+)?')
+@plgn.command('queue add <(?P<what>[-a-zA-Z0-9 `,;!@#$%^&*()_=.{}:"\?\<\>/\[\'\]\\n]+)> *(?P<param>[aA]+)?')
 def queue(data,what,param):
     final = str(what) + " " + str(param)
-    with open(plgn.queued_links,'a') as f:
+    mode = 'a'
+    if not os.access(plgn.queued_links,os.F_OK):
+        mode = 'w' 
+    with open(plgn.queued_links,mode) as f:
         f.write(final)
-        f.write('\n')
+        f.write(',')
     return '{0} added to download queue'.format(str(what))
 
 @plgn.command('begin')
@@ -77,17 +86,22 @@ def begin(data,what = None):
         links = f.read()
     if not os.access(plgn.location,os.F_OK):
         os.mkdir(plgn.location)
-    data = links.split("\n")
+    data = links.split(",")
     data = data[:-1]
     plgn.old=os.listdir(plgn.location)
     output=[]
     for link in data:       
         if not link==" " or not link == "":
             r = link_downloader(link)
-           #logger.debug(link)
-           #logger.debug(r)
             output.append((link,r))
-    os.remove(plgn.queued_links)
+            if r=='1':
+                data.remove(link)
+    if data == []:            
+        os.remove(plgn.queued_links)
+    else:
+        with open(plgn.queued_links,'w') as f:
+            f.write(",".join(data))
+            f.write(",")
     plgn.new = os.listdir(plgn.location)
     final = set(plgn.new) - set(plgn.old) 
     errors = ['Link = '+str(i) + '\n' + 'Error = '+str(j) for i,j in output if j!='1' ]
@@ -101,6 +115,46 @@ def begin(data,what = None):
     else:
         msg = "\n".join(["Following errors were recorded for the given queue","\n".join(errors)])
     return msg
+
+@plgn.command('queue\Z')
+def show(data,what=None):
+    if not os.access(plgn.queued_links,os.F_OK):
+        return "Queue does not exist. Please form a download queue first."
+    with open(plgn.queued_links,'r+') as f:
+        links = f.read()
+    temp = links.split(",")[:-1]
+    final = []
+    data = [str(j) for j in temp if j!='' or j!=' ' or j!='\n']
+    print data
+    if len(temp)==1 and temp[0]=='':
+        return "Please initiate a queue. Either your queue is empty or uninitialized !"
+    else:
+        for i,j in enumerate(data):
+            if j!='' or j!=' ' or j!='\n':
+                i+=1
+                final.append(str(i)+" "+ str(j))
+    return "\n".join(final)
+
+@plgn.command('queue remove (?P<what>[0-9]+)')
+def remove(data,what):
+    if not os.access(plgn.queued_links,os.F_OK):
+        return "Queue does not exist. Please form a download queue first."
+    with open(plgn.queued_links,'r+') as f:
+        links = f.read()
+    data = links.split(",")[:-1]
+    temp =""
+    for i,link in enumerate(data):
+        i+=1
+        if str(what) == str(i):
+            temp = link
+            data.remove(link)
+    
+    with open(plgn.queued_links,'w') as f:
+        f.write(",".join(data))
+        f.write(",")
+    if temp=="":
+        return "Invalid remove execution! Please review the documentation!"
+    return str(temp) + " has been removed from the queue"
 
 @plgn.command('list all downloads')
 def listings(data,what=None):
@@ -134,3 +188,10 @@ def memory(data,what=None):
         if '/dev/root' in line:
             final = line
     return str(final)
+@plgn.command('update youtube downloader')
+def update(data,what=None):
+    proc = subprocess.Popen(['pip','install','--upgrade','youtube_dl'],stdout = subprocess.PIPE)
+    temp = (proc.communicate()[0])
+    logger.debug(str(temp))
+    return str(temp)
+
